# server/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Promuse Backend API
  version: 1.0.0-draft.4
  summary: Authentication and player state for the Promuse mobile rhythm game.
  description: |
    **Phase 1** authentication, **Phase 2** player state, **Phase 3** the economy -
    shop, runs and missions. Each written before its implementation, so the client and
    the server are built against one agreement rather than the client being written
    against whatever the server happened to return.

    Every economy operation is server-authoritative and has its own endpoint rather
    than riding along on a player-state write. The client names what it wants; it never
    names a price, a balance, or a resulting inventory. The only writes that state a
    value outright are the two fields the client legitimately owns: the squad and the
    desktop character.

    Still absent: gacha, and the score and replay half of a run. That endpoint exists -
    POST /v1/runs/{runId}/complete - but carries only a win flag, and Phase 4 adds the
    rest to it rather than introducing a second call that also means "the run ended".

    ## Where the data comes from

    The shapes below mirror `Assets/Arknights/Data/Player/PlayerData.cs`, which is what
    the game holds today as a ScriptableObject that nothing ever writes back. Field
    names are anglicised where the original is Chinese: `reason` (理智) is `stamina`
    here, because "reason" reads as a cause rather than a resource in English.

...

...
      per-device.
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /v1/missions:
    get:
      tags: [missions]
      operationId: getMissionBoards
      summary: Both boards, with progress and what can be claimed.
      description: |
        重置是隐含的 / The reset is implicit rather than scheduled. A board's period key
        is part of the identity of every counter and claim, so a new day is a new key, a
        new key has no rows, and no row reads as zero. There is no job to wipe anything -
        and therefore no job that can fail to run.

        `resetsAt` is when the current board ends, in UTC. The boundary is 04:00 in the
        players' own timezone, which is arithmetic that has no business being on a device
        whose clock is attacker-controlled.
      responses:
        '200':
          description: Both boards.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MissionBoards'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/missions/claims:
    post:
      tags: [missions]
      operationId: claimMission
      summary: Take a finished mission's points.
      description: |
        Pays points, not items. `granted` is empty.

        存在即已领 / Claiming twice is refused by a primary key on
        (account, tab, mission, period) rather than by a check that could race.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClaimMissionRequest'
      responses:
        '200':
          description: Claimed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClaimResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: '`MISSION_NOT_FOUND`.'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description: |
            `MISSION_NOT_COMPLETE`, or `ALREADY_CLAIMED`. The second is not the player's
            mistake - two taps on a slow connection land there - so a client should
            refresh the board rather than say anything alarming.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/missions/reward-claims:
    post:
      tags: [missions]
      operationId: claimMissionReward
      summary: Take a reward the claimed points have unlocked.
      description: |
        The claim row and the items move together - a crash between them would either
        pay twice or pay nothing.

        Thresholds count points that have been CLAIMED, not merely earned.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClaimRewardRequest'
      responses:
        '200':
          description: Claimed, and the items are in the bag.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClaimResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: '`MISSION_NOT_FOUND`.'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description: '`NOT_ENOUGH_POINTS`, or `ALREADY_CLAIMED`.'
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Short-lived access token from any `/v1/auth/*` sign-in. Carries the player id
        and nothing the client is trusted to assert - no level, no balances. Anything
        the server needs in order to decide is read server-side, or it is not a
        decision the server is actually making.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: |
        Client-generated UUID, stable across retries of the *same* logical request and
        different for a new one. The server stores it against the response for 24 hours
        and replays that response rather than acting twice.

        The same key with a *different* body answers 409 with code
        IDEMPOTENCY_KEY_CONFLICT - that is not a retry, and replying with the first
        response would answer a question that was not asked. A missing key answers 400.

        The guarantee is a unique index on (key, endpoint), not a lookup: two
        concurrent retries both read nothing and both proceed, and the database is what
        refuses the second write.
...
  responses:
...
    Unauthorized:
      description: Missing, expired or invalid access token.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    InternalError:
      description: |
        Unexpected server fault. `traceId` correlates with the server log; nothing
        about the internals is ever exposed in `detail`.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Problem:
      type: object
      description: RFC 9457 Problem Details.
      required: [type, title, status, code]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
          description: |
            Stable machine-readable discriminator. The client switches on this, never
            on `title` or `detail`, which are free to be reworded or localised.
          examples: [USERNAME_TAKEN, INVALID_CREDENTIALS, STATE_CONFLICT]
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
        traceId:
          type: string
...
    ServerTime:
      type: string
      format: date-time
      description: |
        Authoritative clock. Every deadline the game enforces is derived from this, so
        that changing the device clock changes nothing.
    PlayerState:
      type: object
      required:
        - playerId
        - displayName
        - level
        - exp
        - expToNextLevel
        - stamina
        - characters
        - squad
        - inventory
        - stateVersion
        - serverTime
      properties:
        playerId:
          type: string
          format: uuid
        displayName:
          type: string
        level:
          type: integer
          minimum: 1
          maximum: 120
          description: Cap matches `PlayerData.GetMaxLevel()`.
        exp:
          type: integer
          minimum: 0
        expToNextLevel:
          type: integer
          description: |
            From `PlayerData.GetMaxExp(level)`. Computed server-side and sent, rather
            than recomputed by the client, so retuning the curve needs no app update
            and the two can never disagree.
        stamina:
          $ref: '#/components/schemas/Stamina'
        characters:
          type: array
          items:
            $ref: '#/components/schemas/CharacterState'
        squad:
          $ref: '#/components/schemas/Squad'
        desktopCharacterId:
          type: [string, 'null']
        inventory:
          type: array
          items:
            $ref: '#/components/schemas/ItemStack'
        stateVersion:
          type: integer
          description: |
            Increments on every accepted write. The same value the `ETag` encodes,
            surfaced in the body so a conflict can be explained to the player rather
            than only handled by the transport.
        serverTime:
          $ref: '#/components/schemas/ServerTime'
    Stamina:
      type: object
      description: |
        `PlayerData.reason` (理智). Regenerates with wall-clock time, so the server owns
        it outright.
      required: [current, max]
      properties:
        current:
          type: integer
          minimum: 0
        max:
          type: integer
          description: From `PlayerData.GetMaxReason(level)`.
        nextPointAt:
          type: [string, 'null']
          format: date-time
          description: When the next point lands. Null at full.
        fullAt:
          type: [string, 'null']
          format: date-time
    CharacterState:
      type: object
      description: Mirrors `Data.Char.CharData`.
      required: [characterId, elite, level, exp, trust]
      properties:
        characterId:
          type: string
          examples: [AMIYA, NOVA, ECHO, PULSE]
          description: |
            Matches the ids under `Resources/Meta/Char`. Only the player's progression
            lives here - artwork, rarity and stats stay client-side content, so a
            balance change is a content update and not a database migration.
        elite:
          type: integer
          minimum: 0
        level:
          type: integer
          minimum: 1
        exp:
          type: integer
          minimum: 0
        trust:
          type: integer
          minimum: 0
    Squad:
      type: array
      minItems: 4
      maxItems: 4
      description: |
        Always four entries, `null` for empty - the shape `PlayerData.GetSquad()`
        normalises to. Fixed length rather than a list of filled slots, because the
        slot position is meaningful.
      items:
        type: [string, 'null']
    ItemStack:
      type: object
      description: Mirrors `Data.Item.ItemStack`.
      required: [itemId, amount]
      properties:
        itemId:
          type: integer
          description: Ids from `Resources/Meta/Item` - 1 Orundum, 2 LMD, 6 Orirock.
        amount:
          type: integer
          minimum: 1
...
    MissionTab:
      type: string
      enum: [Daily, Weekly]
      description: |
        枚举用名字 / Serialised as the name, not an ordinal. An ordinal silently remaps
        every stored and in-flight value when someone reorders the enum, and these names
        are also written into the period keys held per player.
    MissionGoal:
      type: string
      enum: [PlaySong, BuyShopItem]
      description: |
        What a mission counts. A closed set on purpose: each value needs a server-side
        place that advances it, and a goal nothing advances is a mission nobody can
        finish. `PlaySong` advances on a WON run completion; `BuyShopItem` on a checkout,
        once per checkout rather than once per unit.
    MissionState:
      type: object
      required: [missionId, description, goal, target, progress, points, isComplete, isClaimed]
      properties:
        missionId:
          type: string
        description:
          type: string
        goal:
          $ref: '#/components/schemas/MissionGoal'
        target:
          type: integer
          minimum: 1
        progress:
          type: integer
          description: Capped at the target, so a bar never overfills.
        points:
          type: integer
          description: What claiming this mission pays into the board total.
        isComplete:
          type: boolean
        isClaimed:
          type: boolean
          description: |
            做完和领了是两回事 / Finished and claimed are different states. An unclaimed
            mission is worth no points yet, which is the whole point of the manual step.
    RewardState:
      type: object
      required: [rewardId, requiredPoints, itemId, amount, isClaimed, isClaimable]
      properties:
        rewardId:
          type: string
        requiredPoints:
          type: integer
          description: Points that must have been CLAIMED, not merely earned.
        itemId:
          type: integer
        amount:
          type: integer
          minimum: 1
        isClaimed:
          type: boolean
        isClaimable:
          type: boolean
          description: Computed server-side, so the client never re-derives the rule.
    MissionBoard:
      type: object
      required: [tab, title, points, resetsAt, missions, rewards]
      properties:
        tab:
          $ref: '#/components/schemas/MissionTab'
        title:
          type: string
        points:
          type: integer
        resetsAt:
          type: string
          format: date-time
          description: |
            When this board empties, in UTC. The boundary is 04:00 in the players' own
            timezone; the client counts down from this rather than working that out on a
            clock the player controls.
        missions:
          type: array
          items:
            $ref: '#/components/schemas/MissionState'
        rewards:
          type: array
          items:
            $ref: '#/components/schemas/RewardState'
    MissionBoards:
      type: object
      required: [daily, weekly, serverTime]
      properties:
        daily:
          $ref: '#/components/schemas/MissionBoard'
        weekly:
          $ref: '#/components/schemas/MissionBoard'
        serverTime:
          $ref: '#/components/schemas/ServerTime'
    ClaimMissionRequest:
      type: object
      required: [tab, missionId]
      properties:
        tab:
          $ref: '#/components/schemas/MissionTab'
        missionId:
          type: string
    ClaimRewardRequest:
      type: object
      required: [tab, rewardId]
      properties:
        tab:
          $ref: '#/components/schemas/MissionTab'
        rewardId:
          type: string
    ClaimResult:
      type: object
      required: [granted, boards, player]
      properties:
        granted:
          type: array
          description: Empty for a mission claim, which pays points rather than items.
          items:
            $ref: '#/components/schemas/ItemStack'
        boards:
          $ref: '#/components/schemas/MissionBoards'
        player:
          $ref: '#/components/schemas/PlayerState'
...